Created init functions, fixed a bug
Created init functions in classr and download files to make it less awkward to initialize it.
Fixed a bug in main where I assumed that all new post IDs will have a value greater than the older IDs. It worked initially but it has been proven to be wrong now.